frontend/src/features/shared/Service-ai-models/face-recognition/runtime/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
insightface==0.7.3
numpy==1.24.4
Pillow==10.0.1
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
onnxruntime==1.16.3
requests==2.31.0
